+++ builder/builder/index.ts
@@ -1,6 +1,0 @@
-export { generateTableOfContents } from "./generateTableOfContents";
-export { generateStatisticsFile } from "./generateStatisticsFile";
-export { generateSectionReadmes } from "./generateSectionReadmes";
-export { generateGlobalIndex } from "./generateGlobalIndex";
-export { createSectionFile } from "./createSectionFile";
-export { staticContentEntityFactory } from "./staticContentEntityFactory";