+++ builder/types.d.ts
@@ -1,3 +1,0 @@
-declare module "marked-mangle";
-declare module "marked-gfm-heading-id";
-declare module "@eit6609/markdown-templates";