+++ builder/ui/addPageBreak.ts
@@ -1,5 +1,0 @@
-export const addPageBreak = (type: string) => {
-    return type === "md"
-    ? "\n\\newpage \n\n"
-    : '<p style="page-break-after: always;"> </p>';
-}