+++ builder/ui/buildHeadline.ts
@@ -1,5 +1,0 @@
-export const buildHeadline = (content, level: number, type: string) => {
-  if (type === "md") return Array(level).fill("#").join("") + " " + content;
-  if (type === "html") return `<h${level}>${content}</h${level}>`;
-  return "";
-};