+++ builder/ui/wrapToHtmlTag.ts
@@ -1,2 +1,0 @@
-export const wrapToHtmlTag = (tag: string, content = "") =>
-  `<${tag}>${content}</${tag}>`;